Encapsulate common HP logic from osu and catch HP calculations
